requirements.txt
openai>=1.14.0
fastapi
uvicorn
python-dotenv
fpdf
python-multipart
pydantic
python-docx
pymupdf
tqdm
pdfminer.six
PyPDF2
jinja2
requests